fix: move phase and other cli mapping to the output package. Fix the hybrid cluster list command to not need the --env-id flag.

+++ internal/cmd/cluster/delete.go
@@ -32,12 +32,6 @@ qcloud cluster delete 7b2ea926-724b-4de2-b73a-8675c42a6ebe --force`,
 		Run: func(s *state.State, cmd *cobra.Command, args []string) error {
 			clusterID := args[0]
 
-			force, _ := cmd.Flags().GetBool("force")
-			if !util.ConfirmAction(force, cmd.ErrOrStderr(), fmt.Sprintf("Are you sure you want to delete cluster %s?", clusterID)) {
-				fmt.Fprintln(cmd.OutOrStdout(), "Aborted.")
-				return nil
-			}
-
 			ctx := cmd.Context()
 			client, err := s.Client(ctx)
 			if err != nil {
@@ -47,6 +41,24 @@ qcloud cluster delete 7b2ea926-724b-4de2-b73a-8675c42a6ebe --force`,
 			accountID, err := s.AccountID()
 			if err != nil {
 				return err
+			}
+
+			resp, err := client.Cluster().GetCluster(ctx, &clusterv1.GetClusterRequest{
+				AccountId: accountID,
+				ClusterId: clusterID,
+			})
+			if err != nil {
+				return fmt.Errorf("failed to get cluster: %w", err)
+			}
+
+			if resp.GetCluster().GetCloudProviderId() == hybridCloudProviderID {
+				return fmt.Errorf("cluster %s is a hybrid cloud cluster; use \"qcloud hybrid cluster delete\" instead", clusterID)
+			}
+
+			force, _ := cmd.Flags().GetBool("force")
+			if !util.ConfirmAction(force, cmd.ErrOrStderr(), fmt.Sprintf("Are you sure you want to delete cluster %s?", clusterID)) {
+				fmt.Fprintln(cmd.OutOrStdout(), "Aborted.")
+				return nil
 			}
 
 			_, err = client.Cluster().DeleteCluster(ctx, &clusterv1.DeleteClusterRequest{
